+++ app/Http/Controllers/Backend/VendorProductController.php
@@ -7,8 +7,11 @@ use App\Http\Controllers\Controller;
 use App\Models\Brand;
 use App\Models\Category;
 use App\Models\ChildCategory;
+use App\Models\FlashSaleItem;
+use App\Models\OrderProduct;
 use App\Models\Product;
 use App\Models\ProductImages;
+use App\Models\ProductVariant;
 use App\Models\VariantDetails;
 use App\Models\SubCategory;
 use Illuminate\Http\Request;
@@ -142,10 +145,15 @@ class VendorProductController extends Controller
         if ($product->vendor_id != Auth::user()->vendor->id) {
             abort(404);
         }
+        if (OrderProduct::where('product_id', $product->id)->count() > 0) {
+            return response(['status' => 'error', 'message' => 'This product belongs to an order, you can archive only ']);
+        }
+        //delete form flash sale
+        FlashSaleItem::where('product_id', $product->id)->delete();
         //delete product images
         ProductImages::where('product_key', $product->product_key)->delete();
         //delete variant details
-        $variants = VariantDetails::where('product_id', $product->id)->delete();
+        ProductVariant::where('product_id', $product->id)->delete();
         //delete product itself
         $product->delete();
         return response(['status' => 'success', 'message' => 'Deleted successfully']);
